+++ src/oldnews/commands/main.py
@@ -71,6 +71,13 @@ class OpenHomePage(Command):
 
 
 ##############################################################################
+class MarkRead(Command):
+    """Mark the current article as read"""
+
+    BINDING_KEY = "r"
+
+
+##############################################################################
 class MarkAllRead(Command):
     """Mark all unread articles in the current category as read"""