+++ src/oldnews/providers/main.py
@@ -21,6 +21,7 @@ from ..commands import (
     Escape,
     Information,
     MarkAllRead,
+    MarkRead,
     MarkUnread,
     MoveSubscription,
     Next,
@@ -56,6 +57,7 @@ class MainCommands(CommandsProvider):
         yield Help()
         yield from self.maybe(Information)
         yield from self.maybe(MarkAllRead)
+        yield from self.maybe(MarkRead)
         yield from self.maybe(MarkUnread)
         yield from self.maybe(MoveSubscription)
         yield from self.maybe(Next)